Concluído a parte de permissões referente ao acesso e manipulação de telefones

+++ app/Http/Controllers/PhoneController.php
@@ -5,18 +5,43 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Phone;
+use App\Permission;
+use App\GroupPermission;
+use Session;
 
 class PhoneController extends Controller
 {
     public function index($customerId)
     {
-        $phones = Phone::where('customer_id', '=',$customerId)->get();
+        $permissions = [];
+        $groupPermissions = GroupPermission::where('group_id', Session::get('user.group_id'))->get();
         
+        foreach ($groupPermissions as $value) {
+            $permissions[] = (Permission::find($value->permission_id)->toArray())['name'];
+        }
+
+        $phones = [];
+        
+        if (Session::get('user.occupation') == 'admin' || in_array('phone_list', $permissions)) {
+            $phones = Phone::where('customer_id', '=', $customerId)->get();
+        }
+
         return response()->json($phones);
     }
 
     public function store(Request $request)
     {
+        $permissions = [];
+        $groupPermissions = GroupPermission::where('group_id', Session::get('user.group_id'))->get();
+        
+        foreach ($groupPermissions as $value) {
+            $permissions[] = (Permission::find($value->permission_id)->toArray())['name'];
+        }
+
+        if (Session::get('user.occupation') !== 'admin' && !in_array('phone_create', $permissions)) {                   
+            return response()->json(["msg" => "Permissão insuficiente!"], 403);
+        }
+
         $phone = new Phone;
 
         $phone->number = $request->number;
@@ -29,21 +54,43 @@ class PhoneController extends Controller
 
     public function update(Request $request, $id)
     {
+        $permissions = [];
+        $groupPermissions = GroupPermission::where('group_id', Session::get('user.group_id'))->get();
+        
+        foreach ($groupPermissions as $value) {
+            $permissions[] = (Permission::find($value->permission_id)->toArray())['name'];
+        }
+
+        if (Session::get('user.occupation') !== 'admin' && !in_array('phone_edit', $permissions)) {                   
+            return response()->json(["msg" => "Permissão insuficiente!"], 403);
+        }
+
         $phone = Phone::find($id);
 
         $phone->number = $request->number;
 
         $phone->save();
 
-        return response()->json(["msg" =>"ok"]);
+        return response()->json(["msg" => "ok"]);
     }
 
     public function destroy($id)
     {
+        $permissions = [];
+        $groupPermissions = GroupPermission::where('group_id', Session::get('user.group_id'))->get();
+        
+        foreach ($groupPermissions as $value) {
+            $permissions[] = (Permission::find($value->permission_id)->toArray())['name'];
+        }
+
+        if (Session::get('user.occupation') !== 'admin' && !in_array('phone_delete', $permissions)) {                   
+            return response()->json(["msg" => "Permissão insuficiente!"], 403);
+        }
+
         $phone = Phone::find($id);
 
         $phone->delete();
 
-        return response()->json(["msg" =>"ok"]);
+        return response()->json(["msg" => "ok"]);
     }
 }
